app/requirements.txt
flask==2.3.2
flask_sqlalchemy==3.1.1
psycopg2-binary
